[ENH] Put all user interaction on res.partner.relation.all.
