+++ partner_relations/__openerp__.py
@@ -16,7 +16,6 @@
     ],
     "data": [
         "views/res_partner_relation_all.xml",
-        'views/res_partner_relation.xml',
         'views/res_partner.xml',
         'views/res_partner_relation_type.xml',
         'views/menu.xml',
