+++ partner_relations/models/res_partner_relation.py
@@ -8,25 +8,16 @@ class ResPartnerRelation(models.Model):
     """Model res.partner.relation is used to describe all links or relations
     between partners in the database.
 
-    In many parts of the code we have to know whether the active partner is
-    the left partner, or the right partner. If the active partner is the
-    right partner we have to show the inverse name.
-
-    Because the active partner is crucial for the working of partner
-    relationships, we make sure on the res.partner model that the partner id
-    is set in the context where needed.
+    This model is actually only used to store the data. The model
+    res.partner.relation.all, based on a view that contains each record
+    two times, once for the normal relation, once for the inverse relation,
+    will be used to maintain the data.
     """
     _name = 'res.partner.relation'
     _description = 'Partner relation'
     _order = 'left_partner_id, active desc, date_start desc, date_end desc'
 
     allow_self = fields.Boolean(related='type_id.allow_self')
-    any_partner_id = fields.Many2many(
-        comodel_name='res.partner',
-        string='Partner',
-        compute='_compute_any_partner_id',
-        search='_search_any_partner_id'
-    )
     left_partner_id = fields.Many2one(
         comodel_name='res.partner',
         string='Source Partner',
@@ -138,19 +129,6 @@ class ResPartnerRelation(models.Model):
             'type_id': type_domain,
         }
         return {'domain': domain}
-
-    @api.one
-    @api.depends('left_partner_id', 'right_partner_id')
-    def _compute_any_partner_id(self):
-        self.any_partner_id = self.left_partner_id + self.right_partner_id
-
-    @api.model
-    def _search_any_partner_id(self, operator, value):
-        return [
-            '|',
-            ('left_partner_id', operator, value),
-            ('right_partner_id', operator, value),
-        ]
 
     @api.model
     def create(self, vals):
