--- partner_relations/views/res_partner_relation.xml
@@ -1,81 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-
-        <record id="tree_res_partner_relation" model="ir.ui.view">
-            <field name="model">res.partner.relation</field>
-            <field name="arch" type="xml">
-                <tree
-                    string="Partner Relations"
-                    colors="gray:date_end and date_end &lt; current_date or not active;blue:date_start &gt; current_date"
-                    editable="top"
-                    >
-                    <field
-                        name="left_partner_id"
-                        options="{'no_create': True}"
-                        />
-                    <field
-                        name="type_id"
-                        required="True"
-                        options="{'no_create': True}"
-                        />
-                    <field
-                        name="right_partner_id"
-                        options="{'no_create': True}"
-                        />
-                    <field name="date_start" />
-                    <field name="date_end" />
-                    <field name="active" />
-                </tree>
-            </field>
-        </record>
-
-        <record id="search_res_partner_relation" model="ir.ui.view">
-            <field name="model">res.partner.relation</field>
-            <field name="arch" type="xml">
-                <search string="Search Relations">
-                    <field name="any_partner_id" widget="many2one"/>
-                    <field name="left_partner_id"/>
-                    <field name="right_partner_id"/>
-                    <field name="type_id"/>
-                    <field name="active"/>
-                    <group expand="0" string="Group By">
-                        <filter
-                            string="Left Partner"
-                            context="{'group_by': 'left_partner_id'}"
-                            />
-                        <filter
-                            string="Right Partner"
-                            context="{'group_by': 'right_partner_id'}"
-                            />
-                        <filter
-                            string="Relationship Type"
-                            context="{'group_by': 'type_id'}"
-                            />
-                    </group>
-                </search>
-            </field>
-        </record>
-
-        <record
-            id="action_res_partner_relation"
-            model="ir.actions.act_window"
-            >
-            <field name="name">Relations</field>
-            <field name="res_model">res.partner.relation</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree</field>
-            <field name="view_id" ref="tree_res_partner_relation"/>
-            <field name="search_view_id" ref="search_res_partner_relation"/>
-            <field name="domain">[('active', '=', True)]</field>
-            <field name="help" type="html">
-                <p class="oe_view_nocontent_create">
-                    Record and track your partners' relations. Relations may
-                    be linked to other partners with a type either directly
-                    or inversely.
-                </p>
-            </field>
-        </record>
-
-    </data>
-</openerp>
